BETA: Fixing Recommendation bonus to give bonus EXP/SP
git-svn-id: http://svn.l2jserver.com/branches/unstable@5109 d34dfa0f-f91c-0410-86dd-cf06f19dd0ad

+++ L2J_Server_BETA/java/com/l2jserver/gameserver/model/entity/RecoBonus.java
@@ -52,7 +52,7 @@ public final class RecoBonus
 	{
 		double _multiplier = 1;
 		
-		int bonus = getRecoBonus(activeChar);
+		double bonus = getRecoBonus(activeChar);
 		if(bonus > 0)
 			_multiplier = (1 + (bonus / 100));
